--- lib/convert/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "convert"
-version = "0.1.0"
-authors = ["Mr.Panda <xivistudios@gmail.com>"]
-edition = "2018"
-
-[dependencies]
