+++ sdp/src/lib.rs
