+++ stun/Cargo.toml
@@ -12,7 +12,6 @@ md5 = "0.7.0"
 hmac = "0.10.1"
 sha-1 = "0.9.2"
 crc = "1.8.1"
-convert = { path = "../convert" }
 
 [dev-dependencies]
 criterion = {version = "0.3", features = ["html_reports"]}
