+++ turn/Cargo.toml
@@ -17,8 +17,8 @@ description = """
 """
 
 [dependencies]
-stun = { path = "../lib/stun" }
-random_port = { path = "../lib/random_port" }
+stun = { path = "../stun" }
+random_port = { path = "../random_port" }
 tokio = { version = "1", features = ["full"] }
 structopt = { version = "0.3", default-features = false }
 serde = { version = "1.0", features = ["derive"] }
