--- turn/src/state/util.rs
@@ -1,10 +1,0 @@
-// find item first index in vector.
-// pub fn first_index(raw: &Vec<u16>, value: u16) -> Option<usize> {
-//     for (index, item) in raw.iter().enumerate() {
-//         if item == &value {
-//             return Some(index)
-//         }
-//     }
-
-//     None
-// }
